Add files via upload
main: displays use of scope.

We are getting a reference error when we try to log "Hello " and "name" to the console. See if you can move the console.log statement to where it can access the variable name.

We are trying to keep track of the total amount of points that a player has scored, but each time the function is called, our score is being set to zero again. Which statement could you move outside the function so that the total is set to 0 when the page first loads, but every time the function is called the total is incremented by one?